Split up SvgParameters and LayoutParameters for single-line-diagram related code
