+++ diagram-viewer/src/main/java/com/powsybl/diagram/viewer/Model.java
@@ -28,7 +28,7 @@ public class Model {
         this.nadModel = nadModel;
         this.sldModel = sldModel;
         this.nadModel.getSvgParametersBean().bind(showNames);
-        this.sldModel.getSldParametersBean().bind(showNames);
+        this.sldModel.getSvgParametersBean().bind(showNames);
     }
 
     public void setNetwork(Network network) {
