+++ diagram-viewer/src/main/java/com/powsybl/diagram/viewer/sld/SingleLineDiagramViewController.java
@@ -210,22 +210,24 @@ public class SingleLineDiagramViewController extends AbstractDiagramViewControll
                 externCellHeightSpinner.getValueFactory().valueProperty(),
                 internCellHeightSpinner.getValueFactory().valueProperty(),
                 stackHeightSpinner.getValueFactory().valueProperty(),
-                showGridCheckBox.selectedProperty(),
-                showInternalNodesCheckBox.selectedProperty(),
-                drawStraightWiresCheckBox.selectedProperty(),
                 disconnectorsOnBusCheckBox.selectedProperty(),
                 scaleFactorSpinner.getValueFactory().valueProperty(),
-                avoidSVGComponentsDuplicationCheckBox.selectedProperty(),
                 adaptCellHeightToContentCheckBox.selectedProperty(),
                 minSpaceBetweenComponentsSpinner.getValueFactory().valueProperty(),
                 minimumExternCellHeightSpinner.getValueFactory().valueProperty(),
                 busBarAlignmentChoice.valueProperty(),
+                spaceForFeederInfosSpinner.getValueFactory().valueProperty(),
+
+                //SvgParameters
+                showGridCheckBox.selectedProperty(),
+                showInternalNodesCheckBox.selectedProperty(),
+                drawStraightWiresCheckBox.selectedProperty(),
                 centerLabelCheckBox.selectedProperty(),
                 labelDiagonalCheckBox.selectedProperty(),
                 angleLabelSpinner.getValueFactory().valueProperty(),
                 addNodesInfosCheckBox.selectedProperty(),
                 feederInfoSymmetryCheckBox.selectedProperty(),
-                spaceForFeederInfosSpinner.getValueFactory().valueProperty(),
+                avoidSVGComponentsDuplicationCheckBox.selectedProperty(),
                 feederInfosOuterMarginSpinner.getValueFactory().valueProperty(),
                 feederInfosIntraMarginSpinner.getValueFactory().valueProperty()
         );
